Split ConstraintSystem into ProverCS, VerifierCS, and a trait
